add funtionality to buttons
