+++ src/pages/platformAmbient/Trade/TradeCharts/TradeCandleStickChart.tsx
@@ -68,8 +68,6 @@ interface propsIF {
     setShowLatest: Dispatch<SetStateAction<boolean>>;
     updateURL: (changes: updatesIF) => void;
     openMobileSettingsModal: () => void;
-    colorChangeTrigger: boolean;
-    setColorChangeTrigger: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 function TradeCandleStickChart(props: propsIF) {
@@ -1205,8 +1203,6 @@ function TradeCandleStickChart(props: propsIF) {
                             setChartResetStatus={setChartResetStatus}
                             chartResetStatus={chartResetStatus}
                             openMobileSettingsModal={openMobileSettingsModal}
-                            colorChangeTrigger={props.colorChangeTrigger}
-                            setColorChangeTrigger={props.setColorChangeTrigger}
 
                         />
                     </>
